requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Telegram Bot
python-telegram-bot==20.7
telethon==1.34.0

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Google Sheets Integration
gspread==6.0.0
google-auth==2.27.0

# OpenAI API
openai==1.10.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & Utilities
aiofiles==23.2.1
httpx==0.26.0
celery==5.3.6
redis==5.0.1

# Additional
cryptography==41.0.7
